+++ TopTagger/include/TopTaggerUtilities.h
@@ -32,7 +32,7 @@ namespace ttUtility
         const std::vector<TLorentzVector>* jetsLVec_;
         const std::vector<double>* btagFactors_;
         const std::vector<double>* qgLikelihood_;
-        const std::vector<double>* qgMult_;
+        const std::vector<int>* qgMult_;
         const std::vector<double>* qgPtD_;
         const std::vector<double>* qgAxis1_;
         const std::vector<double>* qgAxis2_;
@@ -40,7 +40,7 @@ namespace ttUtility
     public:
         ConstAK4Inputs(const std::vector<TLorentzVector>& jetsLVec, const std::vector<double>& btagFactors, const std::vector<double>& qgLikelihood);
         ConstAK4Inputs(const std::vector<TLorentzVector>& jetsLVec, const std::vector<double>& btagFactors, const std::vector<double>& qgLikelihood, const std::vector<TLorentzVector>& hadGenTops, const std::vector<std::vector<const TLorentzVector*>>& hadGenTopDaughters);
-        void addQGLVectors(const std::vector<double>& qgMult, const std::vector<double>& qgPtD, const std::vector<double>& qgAxis1, const std::vector<double>& qgAxis2);
+        void addQGLVectors(const std::vector<int>& qgMult, const std::vector<double>& qgPtD, const std::vector<double>& qgAxis1, const std::vector<double>& qgAxis2);
         void packageConstituents(std::vector<Constituent>& constituents);
     };
